WIP - scroll direction detection and style changes
