+++ src/components/Banners/MintWheelsBanner/MintWheelsBanner.module.scss
@@ -9,7 +9,8 @@
 	transition: filter var(--animation-time-medium) ease-in-out;
 	transition: transform var(--animation-time-medium) ease-in-out;
 
-	background: none;
+	background: blue;
+
 	padding: 0;
 	text-align: left;
 }
@@ -137,7 +138,7 @@
 
 @media only screen and (min-width: 900px) {
 	.Background {
-		opacity: 0.8;
+		opacity: 1;
 		background-position: 70%;
 	}
 }
