+++ src/components/WilderIcon/WilderIcon.module.scss
@@ -1,22 +1,3 @@
-.FilterBar {
-	position: fixed;
-	top: 29px;
-	margin-left: 0;
-	margin-right: 0;
-
-	display: flex;
-	flex-direction: column;
-	justify-content: space-between;
-
-	background-color: var(--color-background-alt);
-
-	border-radius: var(--box-radius);
-	z-index: 10;
-
-	transition: opacity 0.4s ease-in-out, 0.4s ease-in-out;
-	transform: translateY(0);
-}
-
 .Wilder {
 	position: absolute;
 	top: 0;
@@ -43,38 +24,13 @@
 	height: auto;
 	margin-left: 1px;
 
+	// are these styles used?
 	user-drag: none;
 	-webkit-user-drag: none;
 	user-select: none;
 	-moz-user-select: none;
 	-webkit-user-select: none;
 	-ms-user-select: none;
-}
-
-.Hidden {
-	opacity: 0;
-	transform: translateY(-100px);
-}
-
-.FilterBar > ul {
-	display: flex;
-
-	width: 100%;
-
-	padding: 0 20px 0 20px;
-}
-
-.FilterBar > ul button {
-	font-size: 14px;
-}
-
-.FilterBar > ul > li:not(:first-child) {
-	margin-left: 28px;
-}
-
-.FilterBar > ul button {
-	padding-left: 0;
-	padding-right: 0;
 }
 
 @media only screen and (max-width: 1282px) {
