+++ src/components/WilderIcon/WilderIcon.tsx
@@ -1,0 +1,26 @@
+//- React Imports
+import React from 'react';
+import { useHistory } from 'react-router-dom';
+
+//- Style Imports
+import styles from './WilderIcon.module.scss';
+
+//- Asset Imports
+import wilderIcon from './assets/WWLogo_SVG.svg';
+
+const WilderIcon: React.FC = () => {
+	//- Hooks
+	const history = useHistory();
+
+	const home = () => {
+		history.push('/');
+	};
+
+	return (
+		<div className={styles.Wilder}>
+			<img alt="home icon" src={wilderIcon} onClick={home} />
+		</div>
+	);
+};
+
+export default WilderIcon;
