+++ src/components/WilderIcon/assets/wilder-icon.png
