+++ src/containers/flows/MintWheels/index.tsx
@@ -410,14 +410,12 @@ const MintWheelsFlowContainer = () => {
 					/>
 				</Overlay>
 			)}
-			<div style={{ position: 'relative', marginBottom: 16 }}>
-				<MintWheelsBanner
-					title={'Get your ride for the Metaverse '}
-					label={bannerLabel()}
-					buttonText={buttonText()}
-					onClick={openWizard}
-				/>
-			</div>
+			<MintWheelsBanner
+				title={'Get your ride for the Metaverse '}
+				label={bannerLabel()}
+				buttonText={buttonText()}
+				onClick={openWizard}
+			/>
 		</>
 	);
 };
