+++ src/containers/other/PageHeader/PageHeader.module.scss
@@ -1,0 +1,39 @@
+.NavBar {
+	position: fixed;
+	top: 29px;
+	margin-left: 0;
+	margin-right: 0;
+
+	display: flex;
+	flex-direction: column;
+	justify-content: space-between;
+
+	border-radius: var(--box-radius);
+	z-index: 10;
+
+	transition: opacity 0.4s ease-in-out, 0.4s ease-in-out;
+	transform: translateY(0);
+}
+
+.BannerContainer {
+	position: fixed;
+	z-index: 2;
+	width: 1230px;
+}
+
+.ScrollDown {
+	top: 24px;
+	margin-top: 0px;
+	transition: opacity 0.4s ease-in-out, 0.4s ease-in-out;
+}
+
+.ScrollUp {
+	top: 95px;
+	margin-top: 14px;
+	transition: opacity 0.4s ease-in-out, 0.4s ease-in-out;
+}
+
+.Hidden {
+	opacity: 0;
+	transform: translateY(-100px);
+}
